+++ db/migrate/20101122024407_create_project_users.rb
@@ -1,9 +1,9 @@
 class CreateProjectUsers < ActiveRecord::Migration
   def self.up
     up_sql = %Q(CREATE TABLE `project_user` (
-                `id` SMALLINT(5) UNSIGNED NOT NULL AUTO_INCREMENT,           
-                `project_id` SMALLINT(5) UNSIGNED COMMENT 'ID of the project owner',
-                `user_id` SMALLINT(5) UNSIGNED COMMENT 'ID of the project owner',
+                `id` INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,           
+                `project_id` SMALLINT(11) UNSIGNED COMMENT 'ID of the project owner',
+                `user_id` INT(11) UNSIGNED COMMENT 'ID of the project owner',
                 `updated_at` TIMESTAMP NULL COMMENT 'Datetime of latest update',
                 `created_at` TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Datetime of creation.',
                 PRIMARY KEY  (`id`)
